Revert "get rid of WriteAccountMapEntry instance (#19093)"
This reverts commit f11d938133315a3f7d1258e35b9128650dadd18b.

+++ runtime/src/accounts_index.rs
@@ -284,6 +284,37 @@ impl<T: IndexValue> PreAllocatedAccountMapEntry<T> {
     }
 }
 
+#[self_referencing]
+pub struct WriteAccountMapEntry<T: IndexValue> {
+    owned_entry: AccountMapEntry<T>,
+    #[borrows(owned_entry)]
+    #[covariant]
+    slot_list_guard: RwLockWriteGuard<'this, SlotList<T>>,
+}
+
+impl<T: IndexValue> WriteAccountMapEntry<T> {
+    pub fn from_account_map_entry(account_map_entry: AccountMapEntry<T>) -> Self {
+        WriteAccountMapEntryBuilder {
+            owned_entry: account_map_entry,
+            slot_list_guard_builder: |lock| lock.slot_list.write().unwrap(),
+        }
+        .build()
+    }
+
+    pub fn slot_list(&self) -> &SlotList<T> {
+        &*self.borrow_slot_list_guard()
+    }
+
+    pub fn slot_list_mut<RT>(
+        &mut self,
+        user: impl for<'this> FnOnce(&mut RwLockWriteGuard<'this, SlotList<T>>) -> RT,
+    ) -> RT {
+        let result = self.with_slot_list_guard_mut(user);
+        self.borrow_owned_entry().set_dirty(true);
+        result
+    }
+}
+
 #[derive(Debug, Default, AbiExample, Clone)]
 pub struct RollingBitField {
     max_width: u64,
@@ -1168,20 +1199,12 @@ impl<T: IndexValue> AccountsIndex<T> {
             .map(ReadAccountMapEntry::from_account_map_entry)
     }
 
-    fn slot_list_mut<RT>(
-        &self,
-        pubkey: &Pubkey,
-        user: impl for<'a> FnOnce(&mut RwLockWriteGuard<'a, SlotList<T>>) -> RT,
-    ) -> Option<RT> {
-        let read_lock = self.account_maps[self.bin_calculator.bin_from_pubkey(pubkey)]
+    fn get_account_write_entry(&self, pubkey: &Pubkey) -> Option<WriteAccountMapEntry<T>> {
+        self.account_maps[self.bin_calculator.bin_from_pubkey(pubkey)]
             .read()
-            .unwrap();
-        let get = read_lock.get(pubkey);
-        get.map(|entry| {
-            let result = user(&mut entry.slot_list.write().unwrap());
-            entry.set_dirty(true);
-            result
-        })
+            .unwrap()
+            .get(pubkey)
+            .map(WriteAccountMapEntry::from_account_map_entry)
     }
 
     pub fn handle_dead_keys(
@@ -1319,19 +1342,22 @@ impl<T: IndexValue> AccountsIndex<T> {
     where
         C: Contains<'a, Slot>,
     {
-        self.slot_list_mut(pubkey, |slot_list| {
-            slot_list.retain(|(slot, item)| {
-                let should_purge = slots_to_purge.contains(slot);
-                if should_purge {
-                    reclaims.push((*slot, *item));
-                    false
-                } else {
-                    true
-                }
-            });
-            slot_list.is_empty()
-        })
-        .unwrap_or(true)
+        if let Some(mut write_account_map_entry) = self.get_account_write_entry(pubkey) {
+            write_account_map_entry.slot_list_mut(|slot_list| {
+                slot_list.retain(|(slot, item)| {
+                    let should_purge = slots_to_purge.contains(slot);
+                    if should_purge {
+                        reclaims.push((*slot, *item));
+                        false
+                    } else {
+                        true
+                    }
+                });
+                slot_list.is_empty()
+            })
+        } else {
+            true
+        }
     }
 
     pub fn min_ongoing_scan_root(&self) -> Option<Slot> {
@@ -1693,10 +1719,12 @@ impl<T: IndexValue> AccountsIndex<T> {
         max_clean_root: Option<Slot>,
     ) {
         let mut is_slot_list_empty = false;
-        self.slot_list_mut(pubkey, |slot_list| {
-            self.purge_older_root_entries(slot_list, reclaims, max_clean_root);
-            is_slot_list_empty = slot_list.is_empty();
-        });
+        if let Some(mut locked_entry) = self.get_account_write_entry(pubkey) {
+            locked_entry.slot_list_mut(|slot_list| {
+                self.purge_older_root_entries(slot_list, reclaims, max_clean_root);
+                is_slot_list_empty = slot_list.is_empty();
+            });
+        }
 
         // If the slot list is empty, remove the pubkey from `account_maps`.  Make sure to grab the
         // lock and double check the slot list is still empty, because another writer could have
@@ -1882,12 +1910,12 @@ impl<T: IndexValue> AccountsIndex<T> {
     // if this account has no more entries. Note this does not update the secondary
     // indexes!
     pub fn purge_roots(&self, pubkey: &Pubkey) -> (SlotList<T>, bool) {
-        self.slot_list_mut(pubkey, |slot_list| {
+        let mut write_account_map_entry = self.get_account_write_entry(pubkey).unwrap();
+        write_account_map_entry.slot_list_mut(|slot_list| {
             let reclaims = self.get_rooted_entries(slot_list, None);
             slot_list.retain(|(slot, _)| !self.is_root(*slot));
             (reclaims, slot_list.is_empty())
         })
-        .unwrap()
     }
 }
 
@@ -3894,7 +3922,10 @@ pub mod tests {
 
         secondary_indexes.keys = None;
 
-        index.slot_list_mut(&account_key, |slot_list| slot_list.clear());
+        index
+            .get_account_write_entry(&account_key)
+            .unwrap()
+            .slot_list_mut(|slot_list| slot_list.clear());
 
         // Everything should be deleted
         index.handle_dead_keys(&[&account_key], &secondary_indexes);
@@ -3997,9 +4028,12 @@ pub mod tests {
         // was outdated by the update in the later slot, the primary account key is still alive,
         // so both secondary keys will still be kept alive.
         index.add_root(later_slot, false);
-        index.slot_list_mut(&account_key, |slot_list| {
-            index.purge_older_root_entries(slot_list, &mut vec![], None)
-        });
+        index
+            .get_account_write_entry(&account_key)
+            .unwrap()
+            .slot_list_mut(|slot_list| {
+                index.purge_older_root_entries(slot_list, &mut vec![], None)
+            });
 
         check_secondary_index_mapping_correct(
             secondary_index,
